src/main/java/com/magic/invoicegenerator/service/InvoiceGenerator.java
package com.magic.invoicegenerator.service;

import com.magic.invoicegenerator.entity.Ride;

import java.util.Arrays;

public class InvoiceGenerator {
    private static final double COST_PER_KM = 10;
    private static final int COST_PER_MINUTE = 1;
    private static final int MINIMUM_FARE = 5;

    public double calculateFare(double distance, int time) {
        return Math.max(MINIMUM_FARE, distance * COST_PER_KM + time * COST_PER_MINUTE);
    }

    public double calculateFare(Ride[] rides) {
        double totalFare = 0;
        for (Ride ride : rides) {
            totalFare += calculateFare(ride.distance, ride.time);
        }
        return totalFare;
    }
}
